name changed. Optimized code
